+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/aware/IntelliJPlatformVersionAware.kt
@@ -47,6 +47,14 @@ interface IntelliJPlatformVersionAware {
     val intelliJPlatformPluginConfiguration: ConfigurableFileCollection
 
     /**
+     * Holds the [Configurations.INTELLIJ_PLATFORM_TEST_RUNTIME_FIX_CLASSPATH] configuration with the custom IntelliJ Platform core dependency required for running tests..
+     * It should not be directly accessed.
+     */
+    @get:InputFiles
+    @get:PathSensitive(PathSensitivity.RELATIVE)
+    val intelliJPlatformTestRuntimeFixClasspathConfiguration: ConfigurableFileCollection
+
+    /**
      * Provides access to the IntelliJ Platform dependency artifact path.
      */
     @get:Internal